+++ src/Pages/BlogPage.jsx
@@ -1,10 +1,11 @@
 import React, { useEffect, useMemo, useState } from "react";
 import { motion, AnimatePresence } from "framer-motion";
-import { gsap } from 'gsap'
+import { gsap } from "gsap";
 import { ScrollTrigger } from "gsap/ScrollTrigger";
 import { useGSAP } from "@gsap/react";
+import { useNavigate } from "react-router-dom";
 
-gsap.registerPlugin(ScrollTrigger)
+gsap.registerPlugin(ScrollTrigger);
 
 // LocalStorage keys
 const LS_BLOGS = "ah_blogs_v1";
@@ -151,6 +152,8 @@ function setLS(key, value) {
 const emptyComment = { name: "", text: "" };
 
 export default function BlogHub() {
+  const navigate = useNavigate();
+
   // Data state
   const [blogs, setBlogs] = useState(() => getLS(LS_BLOGS, seedBlogs));
   const [likes, setLikes] = useState(() => getLS(LS_LIKES, {}));
@@ -163,17 +166,7 @@ export default function BlogHub() {
   const [category, setCategory] = useState("");
   const [activeTag, setActiveTag] = useState("");
   const [sort, setSort] = useState("newest");
-  const [showAdd, setShowAdd] = useState(false);
   const [modalBlog, setModalBlog] = useState(null);
-  const [newBlog, setNewBlog] = useState({
-    title: "",
-    excerpt: "",
-    content: "",
-    tags: "",
-    category: "",
-    author: "",
-    cover: "",
-  });
   const [newComment, setNewComment] = useState(emptyComment);
 
   // Persist
@@ -185,18 +178,17 @@ export default function BlogHub() {
 
   // Derived sets
   const allCategories = useMemo(
-    () => Array.from(new Set(blogs.map(b => b.category))).sort(),
+    () => Array.from(new Set(blogs.map((b) => b.category))).sort(),
     [blogs]
   );
   const allTags = useMemo(
-    () => Array.from(new Set(blogs.flatMap(b => b.tags))).sort(),
+    () => Array.from(new Set(blogs.flatMap((b) => b.tags))).sort(),
     [blogs]
   );
 
   // Filtering + sorting
   const visibleBlogs = useMemo(() => {
     let list = blogs.slice();
-
     if (query.trim()) {
       const q = query.toLowerCase();
       list = list.filter(
@@ -210,17 +202,14 @@ export default function BlogHub() {
     }
     if (category) list = list.filter((b) => b.category === category);
     if (activeTag) list = list.filter((b) => b.tags.includes(activeTag));
-
     list.sort((a, b) => {
       if (sort === "likes") {
         const la = likes[a.id] || 0;
         const lb = likes[b.id] || 0;
         if (lb !== la) return lb - la;
       }
-      // newest
       return new Date(b.date) - new Date(a.date);
     });
-
     return list;
   }, [blogs, query, category, activeTag, sort, likes]);
 
@@ -231,12 +220,14 @@ export default function BlogHub() {
     return Array.from(freq.entries()).sort((a, b) => b[1] - a[1]).slice(0, 8);
   }, [blogs]);
 
-  const topLiked = useMemo(() => {
-    return blogs
-      .slice()
-      .sort((a, b) => (likes[b.id] || 0) - (likes[a.id] || 0))
-      .slice(0, 5);
-  }, [blogs, likes]);
+  const topLiked = useMemo(
+    () =>
+      blogs
+        .slice()
+        .sort((a, b) => (likes[b.id] || 0) - (likes[a.id] || 0))
+        .slice(0, 5),
+    [blogs, likes]
+  );
 
   // Actions
   const toggleLike = (id) => {
@@ -287,48 +278,6 @@ export default function BlogHub() {
     });
   };
 
-  const submitBlog = (e) => {
-    e.preventDefault();
-    const id = (blogs.reduce((m, b) => Math.max(m, b.id), 0) || 0) + 1;
-    const now = new Date().toISOString().slice(0, 10);
-    const tags = newBlog.tags
-      .split(",")
-      .map((t) => t.trim())
-      .filter(Boolean);
-
-    const authorName = newBlog.author.trim() || "Guest Author";
-    const item = {
-      id,
-      title: newBlog.title.trim(),
-      excerpt: newBlog.excerpt.trim(),
-      content: newBlog.content.trim(),
-      tags: tags.length ? tags : ["Community"],
-      category: newBlog.category.trim() || "Community",
-      author: {
-        name: authorName,
-        avatar: `https://api.dicebear.com/8.x/initials/svg?seed=${encodeURIComponent(
-          authorName
-        )}`,
-      },
-      date: now,
-      readTime: Math.max(2, Math.round(newBlog.content.split(/\s+/).length / 180)),
-      cover:
-        newBlog.cover.trim() ||
-        "https://images.unsplash.com/photo-1522075469751-3a6694fb2f61?q=80&w=1400&auto=format&fit=crop",
-    };
-    setBlogs((prev) => [item, ...prev]);
-    setShowAdd(false);
-    setNewBlog({
-      title: "",
-      excerpt: "",
-      content: "",
-      tags: "",
-      category: "",
-      author: "",
-      cover: "",
-    });
-  };
-
   // Anim
   const fadeUp = {
     hidden: { opacity: 0, y: 14 },
@@ -336,32 +285,30 @@ export default function BlogHub() {
   };
 
   const bounceHover = {
-    onMouseEnter: e => e.currentTarget.style.transform = "scale(1.07)",
-    onMouseLeave: e => e.currentTarget.style.transform = "scale(1)",
-    style: { transition: "transform 0.3s cubic-bezier(.34,1.56,.64,1)" }
+    onMouseEnter: (e) => (e.currentTarget.style.transform = "scale(1.07)"),
+    onMouseLeave: (e) => (e.currentTarget.style.transform = "scale(1)"),
+    style: { transition: "transform 0.3s cubic-bezier(.34,1.56,.64,1)" },
   };
 
-  useGSAP(()=>{
-    gsap.utils.toArray('.blogCard').forEach((el)=>{
-      gsap.fromTo(el,
+  useGSAP(() => {
+    gsap.utils.toArray(".blogCard").forEach((el) => {
+      gsap.fromTo(
+        el,
+        { y: 30, opacity: 0 },
         {
-          y:30,
-          opacity:0,
-        },
-        {
-          y:0,
-          opacity:1,
-          ease:'bounce.inOut',
-          scrollTrigger:{
-            trigger:el,
-            scrub:1,
-            start:"top 80%",
-            end: 'top 60%',
-          }
+          y: 0,
+            opacity: 1,
+          ease: "bounce.inOut",
+          scrollTrigger: {
+            trigger: el,
+            scrub: 1,
+            start: "top 80%",
+            end: "top 60%",
+          },
         }
-      )
-    })
-  },[])
+      );
+    });
+  }, []);
 
   return (
     <motion.div
@@ -371,14 +318,9 @@ export default function BlogHub() {
       className="min-h-screen w-full py-10 px-4 bg-gradient-to-br from-slate-50 to-blue-50 dark:from-slate-900 dark:to-slate-800 text-slate-900 dark:text-slate-100"
     >
       <div className="max-w-7xl mx-auto space-y-10">
-        {/* Add Blog Button Container - Top Right */}
         <div className="flex justify-end">
           <button
-            onClick={() => setShowAdd(true)}
-            // Mobile sizing
-            // Tablet sizing
-            // Desktop sizing
-            // Large screen sizing
+            onClick={() => navigate("/blogs/add")}
             className="group bg-gradient-to-r from-blue-600 via-purple-600 to-blue-700 text-white rounded-full shadow-2xl hover:shadow-purple-500/25 transition-all duration-500 hover:scale-110 overflow-hidden border-2 border-white/20
             px-4 py-2.5 text-sm
             sm:px-5 sm:py-3 sm:text-base
@@ -386,13 +328,8 @@ export default function BlogHub() {
             lg:px-6 lg:py-3"
             style={{ boxShadow: "0 8px 32px rgba(59, 130, 246, 0.3)" }}
           >
-            {/* Background Animation */}
             <div className="absolute inset-0 bg-gradient-to-r from-purple-600 via-blue-600 to-purple-700 opacity-0 group-hover:opacity-100 transition-opacity duration-500 rounded-full"></div>
-            
-            {/* Shine Effect */}
             <div className="absolute inset-0 -translate-x-full group-hover:translate-x-full transition-transform duration-1000 bg-gradient-to-r from-transparent via-white/20 to-transparent rounded-full"></div>
-            
-            {/* Button Content */}
             <span className="relative z-10 flex items-center gap-1 sm:gap-2">
               <span className="group-hover:rotate-12 transition-transform duration-300 text-sm sm:text-base">➕</span>
               <span>Add Blog</span>
@@ -400,7 +337,6 @@ export default function BlogHub() {
           </button>
         </div>
 
-        {/* Header */}
         <motion.section
           variants={fadeUp}
           initial="hidden"
@@ -416,8 +352,6 @@ export default function BlogHub() {
           <p {...bounceHover} className="text-lg text-slate-600 dark:text-slate-300 max-w-3xl mx-auto">
             Tutorials, showcases, and behind-the-scenes stories from the animation community.
           </p>
-
-          {/* Search & Reset Row */}
           <div className="flex flex-wrap sm:flex-nowrap gap-3 justify-center mt-4">
             <div className="relative w-96">
               <input
@@ -449,8 +383,6 @@ export default function BlogHub() {
               </button>
             </div>
           </div>
-
-          {/* Filter Selects Row */}
           <div className="grid grid-cols-2 sm:grid-cols-3 gap-3 justify-center mt-3 max-w-md mx-auto">
             <select
               value={category}
@@ -488,25 +420,21 @@ export default function BlogHub() {
           </div>
         </motion.section>
 
-        {/* Enhanced Sidebar Widgets Row */}
         <div className="pt-6">
           <motion.h2
             initial={{ opacity: 0, y: -20 }}
             animate={{ opacity: 1, y: 0 }}
-            transition={{ duration: 0.7, ease: 'easeOut' }}
+            transition={{ duration: 0.7, ease: "easeOut" }}
             className="blogCard py-10 text-2xl md:text-4xl font-bold text-center mb-6 bg-gradient-to-r from-blue-600 to-purple-600 bg-clip-text text-transparent"
           >
             Explore Widgets
           </motion.h2>
           <div className="flex flex-col md:flex-row gap-6 justify-center items-stretch mt-8">
-            {/* Top Liked */}
             <div className="blogCard flex-1 min-w-[260px] rounded-3xl border-2 border-gradient-to-r from-blue-200 to-blue-300 dark:from-blue-700 to-blue-600 bg-gradient-to-br from-white/90 via-blue-50/50 to-white/90 dark:from-slate-900/90 dark:via-blue-900/20 dark:to-slate-900/90 shadow-2xl backdrop-blur-xl p-6 flex flex-col relative overflow-hidden">
-              {/* Background Pattern */}
               <div className="absolute inset-0 opacity-5 dark:opacity-10">
                 <div className="absolute top-0 right-0 w-32 h-32 bg-blue-400 rounded-full -translate-y-16 translate-x-16"></div>
                 <div className="absolute bottom-0 left-0 w-24 h-24 bg-blue-300 rounded-full translate-y-12 -translate-x-12"></div>
               </div>
-              
               <div className="relative z-10">
                 <div className="flex items-center gap-3 mb-6" {...bounceHover}>
                   <div className="w-12 h-12 bg-gradient-to-r from-blue-500 to-blue-600 rounded-2xl flex items-center justify-center shadow-lg">
@@ -517,18 +445,22 @@ export default function BlogHub() {
                     <p className="text-sm text-slate-600 dark:text-slate-400">Most popular posts</p>
                   </div>
                 </div>
-                
                 <ul className="space-y-4 flex-1">
                   {topLiked.map((b, index) => (
                     <li key={b.id} className="group">
                       <div className="flex items-center gap-3 p-3 rounded-2xl bg-white/60 dark:bg-slate-800/60 hover:bg-white/80 dark:hover:bg-slate-800/80 transition-all duration-300 hover:scale-105 hover:shadow-lg border border-white/20 dark:border-slate-700/20">
                         <div className="flex-shrink-0">
-                          <div className={`w-8 h-8 rounded-full flex items-center justify-center text-white font-bold text-sm ${
-                            index === 0 ? 'bg-gradient-to-r from-yellow-400 to-yellow-500' :
-                            index === 1 ? 'bg-gradient-to-r from-gray-300 to-gray-400' :
-                            index === 2 ? 'bg-gradient-to-r from-amber-600 to-amber-700' :
-                            'bg-gradient-to-r from-blue-500 to-blue-600'
-                          }`}>
+                          <div
+                            className={`w-8 h-8 rounded-full flex items-center justify-center text-white font-bold text-sm ${
+                              index === 0
+                                ? "bg-gradient-to-r from-yellow-400 to-yellow-500"
+                                : index === 1
+                                ? "bg-gradient-to-r from-gray-300 to-gray-400"
+                                : index === 2
+                                ? "bg-gradient-to-r from-amber-600 to-amber-700"
+                                : "bg-gradient-to-r from-blue-500 to-blue-600"
+                            }`}
+                          >
                             {index + 1}
                           </div>
                         </div>
@@ -552,14 +484,11 @@ export default function BlogHub() {
               </div>
             </div>
 
-            {/* Trending Tags */}
             <div className="blogCard flex-1 min-w-[260px] rounded-3xl border-2 border-gradient-to-r from-purple-200 to-purple-300 dark:from-purple-700 to-purple-600 bg-gradient-to-br from-white/90 via-purple-50/50 to-white/90 dark:from-slate-900/90 dark:via-purple-900/20 dark:to-slate-900/90 shadow-2xl backdrop-blur-xl p-6 flex flex-col relative overflow-hidden">
-              {/* Background Pattern */}
               <div className="absolute inset-0 opacity-5 dark:opacity-10">
                 <div className="absolute top-0 left-0 w-28 h-28 bg-purple-400 rounded-full -translate-y-14 -translate-x-14"></div>
                 <div className="absolute bottom-0 right-0 w-20 h-20 bg-purple-300 rounded-full translate-y-10 translate-x-10"></div>
               </div>
-              
               <div className="relative z-10">
                 <div className="flex items-center gap-3 mb-6" {...bounceHover}>
                   <div className="w-12 h-12 bg-gradient-to-r from-purple-500 to-purple-600 rounded-2xl flex items-center justify-center shadow-lg">
@@ -570,7 +499,6 @@ export default function BlogHub() {
                     <p className="text-sm text-slate-600 dark:text-slate-400">Hot topics now</p>
                   </div>
                 </div>
-                
                 <div className="flex flex-wrap gap-3 flex-1">
                   {trendingTags.map(([t, n]) => (
                     <button
@@ -596,14 +524,11 @@ export default function BlogHub() {
               </div>
             </div>
 
-            {/* My Bookmarks */}
             <div className="blogCard flex-1 min-w-[260px] rounded-3xl border-2 border-gradient-to-r from-yellow-200 to-yellow-300 dark:from-yellow-700 to-yellow-600 bg-gradient-to-br from-white/90 via-yellow-50/50 to-white/90 dark:from-slate-900/90 dark:via-yellow-900/20 dark:to-slate-900/90 shadow-2xl backdrop-blur-xl p-6 flex flex-col relative overflow-hidden">
-              {/* Background Pattern */}
               <div className="absolute inset-0 opacity-5 dark:opacity-10">
                 <div className="absolute top-0 right-0 w-24 h-24 bg-yellow-400 rounded-full -translate-y-12 translate-x-12"></div>
                 <div className="absolute bottom-0 left-0 w-32 h-32 bg-yellow-300 rounded-full translate-y-16 -translate-x-16"></div>
               </div>
-              
               <div className="relative z-10">
                 <div className="flex items-center gap-3 mb-6" {...bounceHover}>
                   <div className="w-12 h-12 bg-gradient-to-r from-yellow-500 to-yellow-600 rounded-2xl flex items-center justify-center shadow-lg">
@@ -614,7 +539,6 @@ export default function BlogHub() {
                     <p className="text-sm text-slate-600 dark:text-slate-400">Saved for later</p>
                   </div>
                 </div>
-                
                 {blogs.filter((b) => bookmarks.has(b.id)).length ? (
                   <ul className="space-y-4 flex-1">
                     {blogs
@@ -658,9 +582,7 @@ export default function BlogHub() {
             </div>
           </div>
         </div>
-        
 
-        {/* Visual Separator */}
         <div className="flex items-center justify-center my-16">
           <div className="flex items-center gap-4">
             <div className="w-16 h-px bg-gradient-to-r from-transparent via-slate-300 dark:via-slate-600 to-transparent"></div>
@@ -669,27 +591,33 @@ export default function BlogHub() {
           </div>
         </div>
 
-        {/* Blogs Section Heading */}
         <motion.section
           variants={fadeUp}
           initial="hidden"
           animate="visible"
           className="blogCard text-center mt-20 mb-8"
         >
-          <div {...bounceHover} className="inline-flex items-center justify-center w-16 h-16 bg-gradient-to-r from-blue-500 to-purple-600 rounded-2xl mb-4 shadow-lg">
+          <div
+            {...bounceHover}
+            className="inline-flex items-center justify-center w-16 h-16 bg-gradient-to-r from-blue-500 to-purple-600 rounded-2xl mb-4 shadow-lg"
+          >
             <span className="text-2xl">📚</span>
           </div>
-          <h2 {...bounceHover} className="text-3xl md:text-4xl font-bold bg-gradient-to-r from-blue-600 to-purple-600 bg-clip-text text-transparent mb-3">
+          <h2
+            {...bounceHover}
+            className="text-3xl md:text-4xl font-bold bg-gradient-to-r from-blue-600 to-purple-600 bg-clip-text text-transparent mb-3"
+          >
             Discover Amazing Content
           </h2>
-          <p {...bounceHover} className="text-lg text-slate-600 dark:text-slate-300 max-w-2xl mx-auto">
+          <p
+            {...bounceHover}
+            className="text-lg text-slate-600 dark:text-slate-300 max-w-2xl mx-auto"
+          >
             Explore our curated collection of tutorials, showcases, and insights from the animation community
           </p>
         </motion.section>
 
-        {/* Main Grid */}
         <section className="grid lg:grid-cols-12 gap-8 mt-8">
-          {/* Blog Grid */}
           <div className="lg:col-span-12 grid sm:grid-cols-2 xl:grid-cols-3 gap-6">
             {visibleBlogs.map((b) => (
               <motion.article
@@ -698,26 +626,25 @@ export default function BlogHub() {
                 className="blogCard group rounded-3xl overflow-hidden border border-transparent border-r-4 border-r-transparent hover:border-r-blue-500 dark:hover:border-r-blue-400 bg-gradient-to-br from-white/80 to-blue-50 dark:from-slate-900/80 dark:to-slate-800/80 shadow-lg hover:shadow-2xl transition-all duration-300 backdrop-blur-lg flex flex-col h-full relative"
               >
                 <div className="relative">
-                   <img
-                     src={b.cover}
-                     alt={b.title}
-                     className="h-44 w-full object-cover"
-                     loading="lazy"
-                     style={{ borderTopLeftRadius: '1.5rem', borderTopRightRadius: '1.5rem' }}
-                   />
-                   <button
-                     onClick={() => toggleBookmark(b.id)}
-                     className={`absolute top-3 right-3 rounded-full px-3 py-1 text-sm font-semibold shadow backdrop-blur-lg transition ${
-                       bookmarks.has(b.id)
-                         ? "bg-gradient-to-r from-yellow-400 to-yellow-300 text-black"
-                         : "bg-white/80 dark:bg-slate-900/70"
-                     }`}
-                   >
-                     {bookmarks.has(b.id) ? "🔖 Saved" : "🔖 Save"}
-                   </button>
-                 </div>
-                 <div className="p-5 flex flex-col flex-1 space-y-4">
-                  {/* Author and Meta */}
+                  <img
+                    src={b.cover}
+                    alt={b.title}
+                    className="h-44 w-full object-cover"
+                    loading="lazy"
+                    style={{ borderTopLeftRadius: "1.5rem", borderTopRightRadius: "1.5rem" }}
+                  />
+                  <button
+                    onClick={() => toggleBookmark(b.id)}
+                    className={`absolute top-3 right-3 rounded-full px-3 py-1 text-sm font-semibold shadow backdrop-blur-lg transition ${
+                      bookmarks.has(b.id)
+                        ? "bg-gradient-to-r from-yellow-400 to-yellow-300 text-black"
+                        : "bg-white/80 dark:bg-slate-900/70"
+                    }`}
+                  >
+                    {bookmarks.has(b.id) ? "🔖 Saved" : "🔖 Save"}
+                  </button>
+                </div>
+                <div className="p-5 flex flex-col flex-1 space-y-4">
                   <div className="flex items-center gap-3">
                     <img
                       src={b.author.avatar}
@@ -725,20 +652,20 @@ export default function BlogHub() {
                       className="w-8 h-8 rounded-full"
                     />
                     <div className="flex flex-col text-xs text-slate-500 dark:text-slate-400">
-                      <span className="font-medium text-slate-700 dark:text-slate-200">{b.author.name}</span>
+                      <span className="font-medium text-slate-700 dark:text-slate-200">
+                        {b.author.name}
+                      </span>
                       <span>
                         {new Date(b.date).toLocaleDateString()} • {b.readTime} min read
                       </span>
                     </div>
                   </div>
-
-                  {/* Title & Excerpt */}
                   <div>
                     <h3 className="font-bold text-lg line-clamp-2 mb-1">{b.title}</h3>
-                    <p className="text-sm text-slate-600 dark:text-slate-300 line-clamp-3">{b.excerpt}</p>
+                    <p className="text-sm text-slate-600 dark:text-slate-300 line-clamp-3">
+                      {b.excerpt}
+                    </p>
                   </div>
-
-                  {/* Tags */}
                   <div className="flex flex-wrap gap-2">
                     {b.tags.map((t) => (
                       <button
@@ -754,8 +681,6 @@ export default function BlogHub() {
                       </button>
                     ))}
                   </div>
-
-                  {/* Actions */}
                   <div className="flex items-center gap-3 justify-start">
                     <button
                       onClick={() => toggleLike(b.id)}
@@ -765,57 +690,54 @@ export default function BlogHub() {
                           : "bg-gradient-to-r from-slate-100 to-slate-200 dark:from-slate-700 dark:to-slate-600 text-slate-700 dark:text-slate-200 hover:from-rose-100 hover:to-pink-100 dark:hover:from-rose-900 dark:hover:to-pink-900"
                       }`}
                     >
-                      {/* Shine Effect */}
                       <div className="absolute inset-0 -translate-x-full group-hover:translate-x-full transition-transform duration-700 bg-gradient-to-r from-transparent via-white/20 to-transparent"></div>
-                      
                       <span className="relative z-10 flex items-center gap-1">
-                        <span className={`transition-transform duration-300 ${likedByMe.has(b.id) ? 'group-hover:scale-125' : 'group-hover:scale-110'}`}>
+                        <span
+                          className={`transition-transform duration-300 ${
+                            likedByMe.has(b.id) ? "group-hover:scale-125" : "group-hover:scale-110"
+                          }`}
+                        >
                           ❤️
                         </span>
                         <span className="font-semibold">{likes[b.id] || 0}</span>
                       </span>
                     </button>
-                    
                     <button
                       onClick={() => copyShareLink(b.id)}
                       className="group relative px-3 py-2 rounded-xl bg-gradient-to-r from-slate-100 to-slate-200 dark:from-slate-700 dark:to-slate-600 text-slate-700 dark:text-slate-200 font-medium hover:from-blue-100 hover:to-purple-100 dark:hover:from-blue-900 dark:hover:to-purple-900 transition-all duration-300 hover:scale-105 overflow-hidden"
                       title="Copy share link"
                     >
-                      {/* Shine Effect */}
                       <div className="absolute inset-0 -translate-x-full group-hover:translate-x-full transition-transform duration-700 bg-gradient-to-r from-transparent via-white/20 to-transparent"></div>
-                      
                       <span className="relative z-10 flex items-center gap-1">
-                        <span className="group-hover:rotate-12 transition-transform duration-300">🔗</span>
+                        <span className="group-hover:rotate-12 transition-transform duration-300">
+                          🔗
+                        </span>
                         <span>Share</span>
                       </span>
                     </button>
                   </div>
-
-                  {/* Read More Button - centered and styled */}
                   <div className="mt-6 flex justify-center">
                     <button
                       onClick={() => openModal(b)}
                       className="group relative px-8 py-3 rounded-2xl font-bold bg-gradient-to-r from-blue-600 via-purple-600 to-blue-700 text-white shadow-lg hover:shadow-2xl transition-all duration-500 hover:scale-110 overflow-hidden"
                       style={{ boxShadow: "0 4px 24px rgba(60,60,120,0.15)" }}
                     >
-                      {/* Background Animation */}
                       <div className="absolute inset-0 bg-gradient-to-r from-purple-600 via-blue-600 to-purple-700 opacity-0 group-hover:opacity-100 transition-opacity duration-500"></div>
-                      
-                      {/* Shine Effect */}
                       <div className="absolute inset-0 -translate-x-full group-hover:translate-x-full transition-transform duration-1000 bg-gradient-to-r from-transparent via-white/20 to-transparent"></div>
-                      
-                      {/* Button Content */}
                       <span className="relative z-10 flex items-center gap-2">
-                        <span className="group-hover:rotate-12 transition-transform duration-300">📖</span>
+                        <span className="group-hover:rotate-12 transition-transform duration-300">
+                          📖
+                        </span>
                         <span>Read More</span>
-                        <span className="group-hover:translate-x-1 transition-transform duration-300">→</span>
+                        <span className="group-hover:translate-x-1 transition-transform duration-300">
+                          →
+                        </span>
                       </span>
                     </button>
                   </div>
                 </div>
               </motion.article>
             ))}
-
             {!visibleBlogs.length && (
               <div className="col-span-full text-center text-slate-500 dark:text-slate-400">
                 No blogs match your filters.
@@ -825,186 +747,6 @@ export default function BlogHub() {
         </section>
       </div>
 
-      {/* Add Blog Modal */}
-      <AnimatePresence>
-        {showAdd && (
-          <motion.div
-            className="fixed inset-0 z-50 flex items-center justify-center p-4"
-            initial={{ opacity: 0 }}
-            animate={{ opacity: 1 }}
-            exit={{ opacity: 0 }}
-          >
-            <div
-              className="absolute inset-0 bg-black/60 backdrop-blur-sm"
-              onClick={() => setShowAdd(false)}
-            />
-            <motion.form
-              onSubmit={submitBlog}
-              initial={{ scale: 0.95, opacity: 0, y: 20 }}
-              animate={{ scale: 1, opacity: 1, y: 0 }}
-              exit={{ scale: 0.95, opacity: 0, y: 20 }}
-              className="relative z-10 w-full max-w-3xl max-h-[85vh] my-8 bg-gradient-to-br from-white/95 via-blue-50/50 to-purple-50/50 dark:from-slate-900/95 dark:via-slate-800/50 dark:to-purple-900/20 rounded-3xl shadow-2xl border-4 border-transparent backdrop-blur-xl overflow-hidden"
-            >
-              {/* Enhanced Header */}
-              <div className="bg-gradient-to-r from-blue-600 via-purple-600 to-blue-700 p-4 text-white relative overflow-hidden">
-                {/* Background Pattern */}
-                <div className="absolute inset-0 opacity-20">
-                  <div className="absolute top-0 right-0 w-32 h-32 bg-white rounded-full -translate-y-16 translate-x-16"></div>
-                  <div className="absolute bottom-0 left-0 w-24 h-24 bg-white rounded-full translate-y-12 -translate-x-12"></div>
-                </div>
-                
-                <div className="relative z-10 flex items-center gap-4">
-                  <div className="w-10 h-10 bg-white/20 rounded-2xl flex items-center justify-center backdrop-blur-sm">
-                    <span className="text-xl">✍️</span>
-                  </div>
-                  <div>
-                    <h3 className="text-xl font-bold">Create a New Blog</h3>
-                    <p className="text-blue-100 text-sm">Share your knowledge with the community</p>
-                  </div>
-                </div>
-                
-                <button
-                  type="button"
-                  onClick={() => setShowAdd(false)}
-                  className="absolute top-3 right-3 w-8 h-8 bg-white/20 hover:bg-white/30 rounded-full flex items-center justify-center backdrop-blur-sm transition-all duration-300 hover:scale-110 z-20"
-                  title="Close"
-                >
-                  <span className="text-white text-base font-bold">×</span>
-                </button>
-              </div>
-
-              {/* Enhanced Form Content */}
-              <div className="p-4 space-y-4 overflow-y-auto max-h-[calc(85vh-120px)]">
-                {/* Title and Author Row */}
-                <div className="grid md:grid-cols-2 gap-3">
-                  <div className="space-y-1">
-                    <label className="text-sm font-semibold text-slate-700 dark:text-slate-200 flex items-center gap-2">
-                      <span className="w-2 h-2 bg-blue-500 rounded-full"></span>
-                      Blog Title *
-                    </label>
-                    <input
-                      required
-                      placeholder="Enter your blog title..."
-                      className="w-full px-3 py-2 rounded-xl border-2 border-slate-200 dark:border-slate-600 bg-white dark:bg-slate-800 focus:border-blue-500 focus:ring-4 focus:ring-blue-500/20 transition-all duration-300 placeholder:text-slate-400 dark:placeholder:text-slate-500"
-                      value={newBlog.title}
-                      onChange={(e) => setNewBlog({ ...newBlog, title: e.target.value })}
-                    />
-                  </div>
-                  
-                  <div className="space-y-1">
-                    <label className="text-sm font-semibold text-slate-700 dark:text-slate-200 flex items-center gap-2">
-                      <span className="w-2 h-2 bg-purple-500 rounded-full"></span>
-                      Author Name
-                    </label>
-                    <input
-                      placeholder="Your name (optional)"
-                      className="w-full px-3 py-2 rounded-xl border-2 border-slate-200 dark:border-slate-600 bg-white dark:bg-slate-800 focus:border-purple-500 focus:ring-4 focus:ring-purple-500/20 transition-all duration-300 placeholder:text-slate-400 dark:placeholder:text-slate-500"
-                      value={newBlog.author}
-                      onChange={(e) => setNewBlog({ ...newBlog, author: e.target.value })}
-                    />
-                  </div>
-                </div>
-
-                {/* Category and Tags Row */}
-                <div className="grid md:grid-cols-2 gap-3">
-                  <div className="space-y-1">
-                    <label className="text-sm font-semibold text-slate-700 dark:text-slate-200 flex items-center gap-2">
-                      <span className="w-2 h-2 bg-green-500 rounded-full"></span>
-                      Category
-                    </label>
-                    <input
-                      placeholder="e.g., CSS, GSAP, Three.js..."
-                      className="w-full px-3 py-2 rounded-xl border-2 border-slate-200 dark:border-slate-600 bg-white dark:bg-slate-800 focus:border-green-500 focus:ring-4 focus:ring-green-500/20 transition-all duration-300 placeholder:text-slate-400 dark:placeholder:text-slate-500"
-                      value={newBlog.category}
-                      onChange={(e) => setNewBlog({ ...newBlog, category: e.target.value })}
-                    />
-                  </div>
-                  
-                  <div className="space-y-1">
-                    <label className="text-sm font-semibold text-slate-700 dark:text-slate-200 flex items-center gap-2">
-                      <span className="w-2 h-2 bg-orange-500 rounded-full"></span>
-                      Tags
-                    </label>
-                    <input
-                      className="w-full px-3 py-2 rounded-xl border-2 border-slate-200 dark:border-slate-600 bg-white dark:bg-slate-800 focus:border-orange-500 focus:ring-4 focus:ring-orange-500/20 transition-all duration-300 placeholder:text-slate-400 dark:placeholder:text-slate-500"
-                      value={newBlog.tags}
-                      onChange={(e) => setNewBlog({ ...newBlog, tags: e.target.value })}
-                    />
-                  </div>
-                </div>
-
-                {/* Cover Image */}
-                <div className="space-y-1">
-                  <label className="text-sm font-semibold text-slate-700 dark:text-slate-200 flex items-center gap-2">
-                    <span className="w-2 h-2 bg-pink-500 rounded-full"></span>
-                    Cover Image URL
-                  </label>
-                  <input
-                    placeholder="https://example.com/image.jpg (optional)"
-                    className="w-full px-3 py-2 rounded-xl border-2 border-slate-200 dark:border-slate-600 bg-white dark:bg-slate-800 focus:border-pink-500 focus:ring-4 focus:ring-pink-500/20 transition-all duration-300 placeholder:text-slate-400 dark:placeholder:text-slate-500"
-                    value={newBlog.cover}
-                    onChange={(e) => setNewBlog({ ...newBlog, cover: e.target.value })}
-                  />
-                </div>
-
-                {/* Excerpt */}
-                <div className="space-y-1">
-                  <label className="text-sm font-semibold text-slate-700 dark:text-slate-200 flex items-center gap-2">
-                    <span className="w-2 h-2 bg-indigo-500 rounded-full"></span>
-                    Short Excerpt *
-                  </label>
-                  <textarea
-                    required
-                    placeholder="Write a brief description of your blog..."
-                    rows={2}
-                    className="w-full px-3 py-2 rounded-xl border-2 border-slate-200 dark:border-slate-600 bg-white dark:bg-slate-800 focus:border-indigo-500 focus:ring-4 focus:ring-indigo-500/20 transition-all duration-300 placeholder:text-slate-400 dark:placeholder:text-slate-500 resize-none"
-                    value={newBlog.excerpt}
-                    onChange={(e) => setNewBlog({ ...newBlog, excerpt: e.target.value })}
-                  />
-                </div>
-
-                {/* Content */}
-                <div className="space-y-1">
-                  <label className="text-sm font-semibold text-slate-700 dark:text-slate-200 flex items-center gap-2">
-                    <span className="w-2 h-2 bg-rose-500 rounded-full"></span>
-                    Blog Content *
-                  </label>
-                  <textarea
-                    required
-                    placeholder="Write your full blog content here... (Markdown supported)"
-                    rows={5}
-                    className="w-full px-3 py-2 rounded-xl border-2 border-slate-200 dark:border-slate-600 bg-white dark:bg-slate-800 focus:border-rose-500 focus:ring-4 focus:ring-rose-500/20 transition-all duration-300 placeholder:text-slate-400 dark:placeholder:text-slate-500 resize-none"
-                    value={newBlog.content}
-                    onChange={(e) => setNewBlog({ ...newBlog, content: e.target.value })}
-                  />
-                </div>
-
-                {/* Enhanced Action Buttons */}
-                <div className="flex flex-col sm:flex-row gap-3 justify-end pt-3 border-t border-slate-200 dark:border-slate-700">
-                  <button
-                    type="button"
-                    onClick={() => setShowAdd(false)}
-                    className="px-5 py-2.5 rounded-xl border-2 border-slate-300 dark:border-slate-600 bg-white dark:bg-slate-800 text-slate-700 dark:text-slate-200 font-semibold hover:bg-slate-50 dark:hover:bg-slate-700 transition-all duration-300 hover:scale-105"
-                  >
-                    Cancel
-                  </button>
-                  <button
-                    className="px-6 py-2.5 rounded-xl bg-gradient-to-r from-blue-600 to-purple-600 hover:from-blue-700 hover:to-purple-700 text-white font-bold shadow-lg hover:shadow-xl transition-all duration-300 hover:scale-105 transform"
-                    type="submit"
-                  >
-                    <span className="flex items-center gap-2">
-                      <span>🚀</span>
-                      <span>Publish Blog</span>
-                    </span>
-                  </button>
-                </div>
-              </div>
-            </motion.form>
-          </motion.div>
-        )}
-      </AnimatePresence>
-
-      {/* Read More Modal */}
       <AnimatePresence>
         {modalBlog && (
           <motion.div
@@ -1024,15 +766,11 @@ export default function BlogHub() {
               transition={{ type: "spring", stiffness: 260, damping: 22 }}
               className="relative z-10 w-full max-w-4xl max-h-[90vh] sm:max-h-[85vh] bg-gradient-to-br from-white/95 via-blue-50/50 to-purple-50/50 dark:from-slate-900/95 dark:via-slate-800/50 dark:to-purple-900/20 rounded-3xl overflow-hidden border-4 border-transparent shadow-2xl backdrop-blur-xl"
             >
-              {/* Enhanced Header */}
               <div className="bg-gradient-to-r from-blue-600 via-purple-600 to-blue-700 p-4 text-white relative overflow-hidden">
-                {/* Background Pattern */}
                 <div className="absolute inset-0 opacity-20">
                   <div className="absolute top-0 right-0 w-32 h-32 bg-white rounded-full -translate-y-16 translate-x-16"></div>
                   <div className="absolute bottom-0 left-0 w-24 h-24 bg-white rounded-full translate-y-12 -translate-x-12"></div>
                 </div>
-                
-                {/* Header Content */}
                 <div className="relative z-10 flex items-center justify-between">
                   <div className="flex items-center gap-4">
                     <div className="w-10 h-10 bg-white/20 rounded-2xl flex items-center justify-center backdrop-blur-sm">
@@ -1043,8 +781,6 @@ export default function BlogHub() {
                       <p className="text-blue-100 text-sm">Reading experience</p>
                     </div>
                   </div>
-                  
-                  {/* Enhanced Close Button */}
                   <button
                     onClick={closeModal}
                     className="w-8 h-8 bg-white/20 hover:bg-white/30 rounded-full flex items-center justify-center backdrop-blur-sm transition-all duration-300 hover:scale-110"
@@ -1055,23 +791,17 @@ export default function BlogHub() {
                 </div>
               </div>
 
-              {/* Enhanced Content */}
               <div className="overflow-y-auto max-h-[calc(90vh-120px)] sm:max-h-[calc(85vh-120px)]">
-                {/* Responsive Image */}
                 <div className="relative">
                   <img
                     src={modalBlog.cover}
                     alt={modalBlog.title}
                     className="h-48 sm:h-56 md:h-64 w-full object-cover"
-                    style={{ borderTopLeftRadius: '1.5rem', borderTopRightRadius: '1.5rem' }}
+                    style={{ borderTopLeftRadius: "1.5rem", borderTopRightRadius: "1.5rem" }}
                   />
                 </div>
-
-                {/* Enhanced Content Section */}
                 <div className="p-4 sm:p-6 md:p-8 space-y-6">
-                  {/* Enhanced Header Section */}
                   <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between gap-4 p-4 rounded-2xl bg-gradient-to-r from-slate-50 to-blue-50 dark:from-slate-800 dark:to-slate-700 border border-slate-200 dark:border-slate-600">
-                    {/* Author Info */}
                     <div className="flex items-center gap-3">
                       <img
                         src={modalBlog.author.avatar}
@@ -1087,8 +817,6 @@ export default function BlogHub() {
                         </div>
                       </div>
                     </div>
-
-                    {/* Enhanced Action Buttons */}
                     <div className="flex flex-wrap gap-2">
                       <button
                         onClick={() => toggleLike(modalBlog.id)}
@@ -1098,17 +826,20 @@ export default function BlogHub() {
                             : "bg-gradient-to-r from-slate-100 to-slate-200 dark:from-slate-700 dark:to-slate-600 text-slate-700 dark:text-slate-200 hover:from-rose-100 hover:to-pink-100 dark:hover:from-rose-900 dark:hover:to-pink-900"
                         }`}
                       >
-                        {/* Shine Effect */}
                         <div className="absolute inset-0 -translate-x-full group-hover:translate-x-full transition-transform duration-700 bg-gradient-to-r from-transparent via-white/20 to-transparent"></div>
-                        
                         <span className="relative z-10 flex items-center gap-1">
-                          <span className={`transition-transform duration-300 ${likedByMe.has(modalBlog.id) ? 'group-hover:scale-125' : 'group-hover:scale-110'}`}>
+                          <span
+                            className={`transition-transform duration-300 ${
+                              likedByMe.has(modalBlog.id)
+                                ? "group-hover:scale-125"
+                                : "group-hover:scale-110"
+                            }`}
+                          >
                             ❤️
                           </span>
                           <span className="font-semibold">{likes[modalBlog.id] || 0}</span>
                         </span>
                       </button>
-                      
                       <button
                         onClick={() => toggleBookmark(modalBlog.id)}
                         className={`group relative px-3 py-2 rounded-xl font-medium transition-all duration-300 hover:scale-105 overflow-hidden ${
@@ -1117,32 +848,29 @@ export default function BlogHub() {
                             : "bg-gradient-to-r from-slate-100 to-slate-200 dark:from-slate-700 dark:to-slate-600 text-slate-700 dark:text-slate-200 hover:from-yellow-100 hover:to-yellow-200 dark:hover:from-yellow-900 dark:hover:to-yellow-800"
                         }`}
                       >
-                        {/* Shine Effect */}
                         <div className="absolute inset-0 -translate-x-full group-hover:translate-x-full transition-transform duration-700 bg-gradient-to-r from-transparent via-white/20 to-transparent"></div>
-                        
                         <span className="relative z-10 flex items-center gap-1">
-                          <span className="group-hover:rotate-12 transition-transform duration-300">🔖</span>
+                          <span className="group-hover:rotate-12 transition-transform duration-300">
+                            🔖
+                          </span>
                           <span>{bookmarks.has(modalBlog.id) ? "Saved" : "Save"}</span>
                         </span>
                       </button>
-                      
                       <button
                         onClick={() => copyShareLink(modalBlog.id)}
                         className="group relative px-3 py-2 rounded-xl bg-gradient-to-r from-slate-100 to-slate-200 dark:from-slate-700 dark:to-slate-600 text-slate-700 dark:text-slate-200 font-medium hover:from-blue-100 hover:to-purple-100 dark:hover:from-blue-900 dark:hover:to-purple-900 transition-all duration-300 hover:scale-105 overflow-hidden"
                         title="Copy share link"
                       >
-                        {/* Shine Effect */}
                         <div className="absolute inset-0 -translate-x-full group-hover:translate-x-full transition-transform duration-700 bg-gradient-to-r from-transparent via-white/20 to-transparent"></div>
-                        
                         <span className="relative z-10 flex items-center gap-1">
-                          <span className="group-hover:rotate-12 transition-transform duration-300">🔗</span>
+                          <span className="group-hover:rotate-12 transition-transform duration-300">
+                            🔗
+                          </span>
                           <span>Share</span>
                         </span>
                       </button>
                     </div>
                   </div>
-
-                  {/* Enhanced Title and Tags */}
                   <div className="space-y-4">
                     <h2 className="text-2xl sm:text-3xl font-bold text-slate-800 dark:text-slate-100 leading-tight">
                       {modalBlog.title}
@@ -1158,15 +886,13 @@ export default function BlogHub() {
                       ))}
                     </div>
                   </div>
-
-                  {/* Enhanced Article Content */}
                   <article className="prose dark:prose-invert max-w-none text-base sm:text-lg leading-relaxed">
                     {modalBlog.content.split("\n").map((p, i) => (
-                      <p key={i} className="mb-4 text-slate-700 dark:text-slate-200">{p}</p>
+                      <p key={i} className="mb-4 text-slate-700 dark:text-slate-200">
+                        {p}
+                      </p>
                     ))}
                   </article>
-
-                  {/* Enhanced Comments Section */}
                   <div className="pt-4 border-t border-slate-200 dark:border-slate-700">
                     <h4 className="font-bold text-lg mb-4 text-slate-800 dark:text-slate-100 flex items-center gap-2">
                       <span className="text-xl">💬</span>
@@ -1179,7 +905,9 @@ export default function BlogHub() {
                           className="flex flex-col sm:flex-row sm:items-start sm:justify-between gap-3 p-4 rounded-2xl bg-gradient-to-r from-slate-50 to-blue-50 dark:from-slate-800 dark:to-slate-700 border border-slate-200 dark:border-slate-600"
                         >
                           <div className="flex-1">
-                            <div className="text-sm font-semibold text-slate-800 dark:text-slate-100">{c.name}</div>
+                            <div className="text-sm font-semibold text-slate-800 dark:text-slate-100">
+                              {c.name}
+                            </div>
                             <div className="text-sm text-slate-600 dark:text-slate-300 whitespace-pre-wrap mt-1">
                               {c.text}
                             </div>
@@ -1202,9 +930,10 @@ export default function BlogHub() {
                         </div>
                       )}
                     </div>
-
-                    {/* Enhanced Comment Form */}
-                    <form onSubmit={submitComment} className="grid grid-cols-1 sm:grid-cols-5 gap-3 p-4 rounded-2xl bg-gradient-to-r from-slate-50 to-blue-50 dark:from-slate-800 dark:to-slate-700 border border-slate-200 dark:border-slate-600">
+                    <form
+                      onSubmit={submitComment}
+                      className="grid grid-cols-1 sm:grid-cols-5 gap-3 p-4 rounded-2xl bg-gradient-to-r from-slate-50 to-blue-50 dark:from-slate-800 dark:to-slate-700 border border-slate-200 dark:border-slate-600"
+                    >
                       <input
                         required
                         placeholder="Your name"
